refactor(Code generators): move stuff into folders
