+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/EnumMatch/EnumMatchGeneratorTests.EnumsWithSameNameInDifferentTypes.verified.txt
