+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/EnumMatch/EnumMatchGeneratorTests.GeneratesMatchExtensionClass.verified.txt
