+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/EnumMatch/EnumMatchGeneratorTests.SupportsNestedTypes.verified.txt
