+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Lambdas/LambdaGeneratorConstructorLambdaTest.SupportsClasses.verified.txt
