+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Lambdas/LambdaGeneratorConstructorLambdaTest.SupportsClassesWithPrimaryConstructor.verified.txt
