+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Lambdas/LambdaGeneratorConstructorLambdaTest.SupportsGenericTypes.verified.txt
