+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Lambdas/LambdaGeneratorConstructorLambdaTest.SupportsStructsWithPrimaryConstructor.verified.txt
