+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Lambdas/LambdaGeneratorConstructorLambdaTest.cs
@@ -2,7 +2,7 @@ using FluentAssertions;
 using FxKit.CompilerServices.CodeGenerators.Lambdas;
 using FxKit.CompilerServices.Tests.TestUtils;
 
-namespace FxKit.CompilerServices.Tests.UnitTests.CodeGenerators;
+namespace FxKit.CompilerServices.Tests.UnitTests.CodeGenerators.Lambdas;
 
 public class LambdaGeneratorConstructorLambdaTest
 {
