+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Lambdas/LambdaGeneratorMethodLambdaTests.GeneratesTheLambdaMethod.verified.txt
