+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Transformers/TransformerGeneratorReferencedAssembliesTests.IncludesFunctorsAndBehaviorFromReferencedAssemblies.verified.txt
