+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Transformers/TransformerGeneratorReferencedAssembliesTests.cs
@@ -5,7 +5,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using static FxKit.CompilerServices.Tests.TestUtils.CodeGeneratorTestUtil;
 
-namespace FxKit.CompilerServices.Tests.UnitTests.CodeGenerators;
+namespace FxKit.CompilerServices.Tests.UnitTests.CodeGenerators.Transformers;
 
 public class TransformerGeneratorReferencedAssembliesTests
 {
@@ -56,10 +56,9 @@ public class TransformerGeneratorReferencedAssembliesTests
 
             // Assert that we didn't recompute the output.
             var result = driver.GetRunResult().Results.Single();
-            var outputs = (
-                from step in result.TrackedOutputSteps.Values.SelectMany(x => x)
+            var outputs = from step in result.TrackedOutputSteps.Values.SelectMany(x => x)
                 from outputStep in step.Outputs
-                select outputStep);
+                select outputStep;
             outputs.Should()
                 .AllSatisfy(output => output.Reason.Should().Be(IncrementalStepRunReason.Cached));
 
@@ -119,10 +118,10 @@ public class TransformerGeneratorReferencedAssembliesTests
                         {
                             using System;
                             using FxKit.CompilerServices;
-
+                        
                             [Functor]
                             public struct NewFunctor<T>;
-
+                        
                             public static class NewFunctor
                             {
                                 [GenerateTransformer]
@@ -190,10 +189,10 @@ public class TransformerGeneratorReferencedAssembliesTests
                         using System;
                         using System.Collections.Generic;
                         using FxKit.CompilerServices;
-
+                    
                         [Functor]
                         public record Filtered<T>(IReadOnlyList<T> Items);
-
+                    
                         public static class FilteredExtensions
                         {
                             [GenerateTransformer]
@@ -201,12 +200,12 @@ public class TransformerGeneratorReferencedAssembliesTests
                                 this Filtered<T> source,
                                 Func<T, U> selector) => new Filtered<U>(source.Items.Select(selector).ToList())
                         }
-
+                    
                         namespace Inner
                         {
                             [Functor]
                             public record Paged<T>(IReadOnlyList<T> Items);
-
+                    
                             public static class PagedExtensions
                             {
                                 [GenerateTransformer]
@@ -245,14 +244,14 @@ public class TransformerGeneratorReferencedAssembliesTests
                         using System;
                         using System.Threading.Tasks;
                         using System.Collections.Generic;
-
+                    
                         public static partial class TaskExtensions
                         {
                             public static async Task<U> Map<T, U>(
                                 this Task<T> source,
                                 Func<T, U> selector) => selector(await source);
                         }
-
+                    
                         namespace Collections
                         {
                             public static partial class EnumerableAndListExtensions
@@ -260,7 +259,7 @@ public class TransformerGeneratorReferencedAssembliesTests
                                 public static IEnumerable<U> Map<T, U>(
                                     this IEnumerable<T> source,
                                     Func<T, U> selector) => source.Select(selector);
-
+                    
                                 public static IReadOnlyList<U> Map<T, U>(
                                     this IReadOnlyList<T> source,
                                     Func<T, U> selector) => source.Select(selector).ToList();
@@ -287,10 +286,10 @@ public class TransformerGeneratorReferencedAssembliesTests
                     {
                         using System;
                         using FxKit.CompilerServices;
-
+                    
                         [Functor]
                         public struct Option<T> where T : notnull {}
-
+                    
                         public static partial class Option
                         {
                             [GenerateTransformer]
