+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Transformers/TransformerGeneratorTests.GeneratesTransformers.verified.txt
