+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Unions/UnionGeneratorTests.GeneratesUnion.verified.txt
