+++ src/FxKit.CompilerServices.Tests/UnitTests/CodeGenerators/Unions/UnionGeneratorTests.cs
@@ -1,8 +1,8 @@
 ﻿using FluentAssertions;
-using FxKit.CompilerServices.CodeGenerators;
+using FxKit.CompilerServices.CodeGenerators.Unions;
 using FxKit.CompilerServices.Tests.TestUtils;
 
-namespace FxKit.CompilerServices.Tests.UnitTests.CodeGenerators;
+namespace FxKit.CompilerServices.Tests.UnitTests.CodeGenerators.Unions;
 
 public class UnionGeneratorTests
 {
