+++ src/FxKit.CompilerServices/Analyzers/MustBePartialAnalyzer.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Immutable;
-using FxKit.CompilerServices.CodeGenerators;
 using FxKit.CompilerServices.CodeGenerators.Lambdas;
+using FxKit.CompilerServices.CodeGenerators.Unions;
 using FxKit.CompilerServices.Utilities;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
