+++ src/FxKit.CompilerServices/CodeGenerators/Transformers/TransformerClassBuilder.cs
@@ -1,3 +1,4 @@
+using FxKit.CompilerServices.Utilities;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
